refactor: 移动 ellipsisLabel() 方法的位置
